+++ src/main/java/ru/rolemodel/model/role/RoleModelService.java
@@ -8,5 +8,5 @@ import java.util.List;
 public interface RoleModelService {
     String saveRoles(RoleEntity roleEntity);
     List<RoleEntity> getRoles(String isService);
-    String deleteRole(String idService, Long idRole);
+    String deleteRole(String idService, Integer idRole);
 }
